Journal 0.3
I am still working on naming conventions.

This is an update to the project labeled Journal 3.0 (which is really
more like 0.2). I have replaced the python dictionary used to manage the
journal entries with a custom JObject, which is a dictionary that is
filled with JEntries, which are objects that are more specialized
versions of the dictionary entries that were used before.

The thing that finally pushed me to implement this new approach is the
desire to see better tags management for each entry, incorporating more
buttons in the GUI and less utilization of text entry fields, making for
finer control on the user-end. There will be other changes, later to
come.

All of this is to take better advantage of the OOP structure of Python,
making for cleaner code and a better user experience.