Functional Update
This is a working update. It is more modular and is better to look at.

The graphing function has been implemented with few issues, so all of
the buttons on the GUI now work.

The save functions have been deprecated so that the program now saves a
JEntry every time the the JEntry of the main program updates or the
program is closed.

Fixed a recursive list bug with the tags management system.

Created the tags management module which will eventually be the
superclass for every module that deals with tags.